PInvoke: First version with Kernel32 from some of my projects
The `Developer.md` documentation is based on research to try and consolidate various sources on the Internet of doing P/Invoke. There's a lot of information, but it's mixed and sometimes hard to find. The goal of this repository is to keep concrete implementations that I can reuse in other projects (or add to), as well as document findings from P/Invoke.

Not covered in the documentation very well are classes and not using a `StringBuilder`